pyproject.toml
[project]
name = "snapp"
version = "0.1.0"
description = ""
authors = [
    {name = "amirhosein soofi",email = "amirhosein79wyvern@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-telegram-bot (==20.0)",
    "tinydb (>=4.8.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
